collect_minerals_and_gas: Change worker assignment
Sometimes the same worker get different an order (e.g. build a supply)
and another order (e.g. build a refinery) right on the next iteration.
This happens because we use only idle workers for the build actions and
as the game goes further the number of idle workers become relatively small.

In order to fix it lets split the actions:
Send the idle workers to gather resources, select any available worker
for the build actions.