api: Reimplement UnitPos class
The UnitPos class was vague and worked simultaneously with point (e.g.
center of a unit) and the numpy arrays. As a result some methods were
broken if user provided points instead of arrays.

This commit changes the behaviour: now the very core of UnitPos is
always center of a unit. However, one can create UnitPos using arrays as well.
As a result, the center property was removed cause it is not needed.