collect_minerals_and_gas: Avoid errors during...
refinery placement. Actually, some time passes until an SCV starts to
build a refinery on the selected geyser. If the same order will be
issued again for the same vespene, we will receive a error.

Lets avoid it by removing the location of a geyser from the cached list
before issuing an order. The action can't fail cause we did all the
required checks.